requirements.txt
fitparse
geopandas
notebook
jupyter
descartes
cpython
